requirements.txt
streamlit==1.37.0
ultralytics==8.3.241
opencv-python-headless==4.10.0.84
numpy
pillow
onnxruntime
